+++ Sources/TorusUtils/Extensions/HelperExtension.swift
@@ -1,0 +1,37 @@
+//
+//  File.swift
+//  
+//
+//  Created by Dhruv Jaiswal on 09/04/23.
+//
+
+import Foundation
+
+// Necessary for decryption
+
+extension Sequence where Element == UInt8 {
+    var data: Data { .init(self) }
+    var hexa: String { map { .init(format: "%02x", $0) }.joined() }
+}
+
+extension Data {
+    init?(hexString: String) {
+        let length = hexString.count / 2
+        var data = Data(capacity: length)
+        for i in 0 ..< length {
+            let j = hexString.index(hexString.startIndex, offsetBy: i * 2)
+            let k = hexString.index(j, offsetBy: 2)
+            let bytes = hexString[j ..< k]
+            if var byte = UInt8(bytes, radix: 16) {
+                data.append(&byte, count: 1)
+            } else {
+                return nil
+            }
+        }
+        self = data
+    }
+
+    func addLeading0sForLength64() -> Data {
+        Data(hex: toHexString().addLeading0sForLength64())
+    }
+}
